requirements.txt
streamlit>=1.24.0
newsapi-python>=0.2.7
openai>=1.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
